+++ .github/pull_request_template.md
@@ -1,0 +1,25 @@
+<!-- First of all, THANK YOU for your contribution. -->
+
+## Changes Proposed:
+- 
+- 
+
+## Issues Addressed:
+<!-- If your fix has a relating issue, link it below -->
+- Closes
+
+## SOURCE:
+<!-- If you can, include a source that can strengthen your claim -->
+
+## Tests Performed:
+<!-- Does it build without errors? Did you test in-game? What did you test? On which OS did you test? Describe any other tests performed -->
+- 
+- 
+
+
+## How to Test the Changes:
+<!-- Describe in a detailed step-by-step order how to test the changes -->
+
+1.
+2.
+3.
