--- conf/conf.sh.dist
@@ -1,32 +1,0 @@
-#!/usr/bin/env bash
-
-## CUSTOM SQL - Important file used by the db_assembler.sh
-## Keep only the required variables (base sql files or updates, depending on the DB)
-
-## BASE SQL
-
-DB_AUTH_CUSTOM_PATHS+=(
-    $MOD_QUEST_STATUS_ROOT"/sql/auth/base/"
-)
-
-DB_CHARACTERS_CUSTOM_PATHS+=(
-    $MOD_QUEST_STATUS_ROOT"/sql/characters/base/"
-)
-
-DB_WORLD_CUSTOM_PATHS+=(
-    $MOD_QUEST_STATUS_ROOT"/sql/world/base/"
-)
-
-## UPDATES
-
-DB_AUTH_UPDATES_PATHS+=(
-    $MOD_QUEST_STATUS_ROOT"/sql/auth/updates/"
-)
-
-DB_CHARACTERS_UPDATES_PATHS+=(
-    $MOD_QUEST_STATUS_ROOT"/sql/characters/updates/"
-)
-
-DB_WORLD_UPDATES_PATHS+=(
-    $MOD_QUEST_STATUS_ROOT"/sql/world/updates/"
-)
