+++ src/qs_loader.cpp
